+++ src/main/php/broker/param/OneOfParamBroker.php
@@ -9,6 +9,7 @@
  */
 namespace stubbles\input\broker\param;
 use stubbles\input\valuereader\CommonValueReader;
+use stubbles\lang\exception\RuntimeException;
 use stubbles\lang\reflect\annotation\Annotation;
 /**
  * Filter values based on a @Request[OneOf] annotation.
@@ -31,7 +32,8 @@ class OneOfParamBroker extends MultipleSourceParamBroker
      * reads default value
      *
      * @param   Annotation  $annotation
-     * @return  array
+     * @return  string[]
+     * @throws  RuntimeException
      */
     private function getAllowedValues(Annotation $annotation)
     {
@@ -39,6 +41,6 @@ class OneOfParamBroker extends MultipleSourceParamBroker
             return array_map('trim', explode('|', $annotation->getAllowed()));
         }
 
-        return [];
+        throw new RuntimeException('No list of allowed values in annotation @Request[OneOf]');
     }
 }