+++ src/test/php/broker/param/OneOfParamBrokerTest.php
@@ -165,4 +165,16 @@ class OneOfParamBrokerTest extends MultipleSourceParamBrokerTest
                             )
         );
     }
+
+    /**
+     * @test
+     * @expectedException  stubbles\lang\exception\RuntimeException
+     * @since  3.0.0
+     */
+    public function throwsRuntimeAnnotationWhenListOfAllowedValuesIsMissing()
+    {
+        $this->paramBroker->procure($this->mockRequest(((string) $this->getExpectedValue())),
+                                    $this->createRequestAnnotation()
+        );
+    }
 }